errors: validation multierror proof of concept
When running a `Validation()` method of a model or a message, the first
validation failure is returned. From the API perspective this returns a
random error message.

If an operation failed because of validation failure it would be better
to return all validation issues at once. This helps avoiding
fix-submit-error-change loop.

The possible downside of this solution is that a group of errors must be
represented as a single ABCICode. That might require to hide precise
error codes under a single one.

This is an explanation for #529